Initialize repository for plugin
Over time, I've noticed a lot of design patterns in our agency's website
designs that are really painful to implement in Beaver Builder.
Components requiring overlapping design elements, things requiring
precise positioning, or even simple issues of alignment often have to be
crudely slapped together with code that's external to Beaver Builder,
and then crammed inside of it. This displeases the developer. The
developer must always be pleased.

The intention of this plugin is to create a centralized place for custom
Beaver Builder modules that fit our use cases. This will, hopefully,
please the developer.